feat: add product and category modules with image upload functionality

+++ src/modules/categories/categories.service.ts
@@ -1,6 +1,10 @@
 import { Category, CategoryDocument } from './entities/category.entity';
 import { CreateCategoryDto, UpdateCategoryDto } from './dto/category.dto';
-import { Injectable } from '@nestjs/common';
+import {
+  ConflictException,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
 import { InjectModel } from '@nestjs/mongoose';
 import { Model } from 'mongoose';
 import { SubcategoriesService } from '../subcategories/subcategories.service';
@@ -22,6 +26,18 @@ export class CategoriesService {
   ) {}
 
   async create(createCategoryDto: CreateCategoryDto) {
+    const publicId = `categories/${sanitizedName(createCategoryDto.name)}`;
+
+    const existingCategory = await this.categoryModel.findOne({
+      publicIdImage: publicId,
+    });
+
+    if (existingCategory) {
+      throw new ConflictException(
+        'Já existe uma categoria com esse nome (imagem).',
+      );
+    }
+
     const imageUrl = await this.bucketService.uploadBase64Image(
       createCategoryDto.urlImage,
       'categories',
@@ -31,6 +47,7 @@ export class CategoriesService {
       ...createCategoryDto,
       createdAt: Date.now(),
       urlImage: imageUrl,
+      publicIdImage: publicId,
     });
     return category.save();
   }
@@ -48,12 +65,18 @@ export class CategoriesService {
   }
 
   async remove(id: string) {
-    const subcategories = await this.subcategoryModel.find({ categoryId: id });
+    const category = await this.categoryModel.findById(id);
 
+    if (!category) {
+      throw new NotFoundException('Categoria não encontrada');
+    }
+
+    const subcategories = await this.subcategoryModel.find({ categoryId: id });
     for (const subcategory of subcategories) {
       await this.subcategoriesService.remove(subcategory._id.toString());
     }
 
+    await this.bucketService.deleteImageByName(category.publicIdImage);
     return this.categoryModel.findByIdAndDelete(id);
   }
 }
