+++ src/modules/products/products.service.ts
@@ -1,4 +1,10 @@
-import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
+import {
+  ConflictException,
+  HttpException,
+  HttpStatus,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
 import { Product, ProductDocument } from './entities/product.entity';
 import { InjectModel } from '@nestjs/mongoose';
 import { ProductDto, UpdateProductDto } from './dto/product.dto';
@@ -14,6 +20,18 @@ export class ProductsService {
   ) {}
 
   async create(createProductDto: ProductDto) {
+    const publicId = `products/${sanitizedName(createProductDto.name)}`;
+
+    const existingProduct = await this.productModel.findOne({
+      publicIdImage: publicId,
+    });
+
+    if (existingProduct) {
+      throw new ConflictException(
+        'Já existe um produto com esse nome (imagem).',
+      );
+    }
+
     const imageUrl = await this.bucketService.uploadBase64Image(
       createProductDto.urlImage,
       'products',
@@ -21,8 +39,8 @@ export class ProductsService {
     );
     const product = new this.productModel({
       ...createProductDto,
-      createdAt: Date.now(),
       urlImage: imageUrl,
+      publicIdImage: publicId,
     });
     return product.save();
   }
@@ -63,7 +81,6 @@ export class ProductsService {
     return product;
   }
 
-  // Atualizar um produto
   async update(id: string, updateProductDto: UpdateProductDto) {
     const updatedProduct = await this.productModel
       .findByIdAndUpdate(id, updateProductDto, {
@@ -80,7 +97,14 @@ export class ProductsService {
     return updatedProduct;
   }
 
-  remove(id: string) {
+  async remove(id: string) {
+    const product = await this.productModel.findById(id);
+
+    if (!product) {
+      throw new NotFoundException('Product not found');
+    }
+
+    await this.bucketService.deleteImageByName(product.publicIdImage);
     return this.productModel.findByIdAndDelete(id);
   }
 }
